Split out VerilatorExecutive from VerilatorBackend.
